basic test for extracted form component
