+++ app/react/Metadata/components/specs/MediaModal.spec.tsx
@@ -7,7 +7,7 @@ import { Provider } from 'react-redux';
 
 import { RenderAttachment } from 'app/Attachments/components/RenderAttachment';
 
-import { WebMediaResourceForm } from 'app/Attachments/components/WebMediaResource';
+import { WebMediaResourceForm } from 'app/Attachments/components/WebMediaResourceForm';
 import { MediaModal, MediaModalProps, MediaModalType } from '../MediaModal';
 
 const mockStore = configureMockStore([thunk]);
